Move work item total cost calculation to model
